Updating index.php
Added in the ability to show images. You need to make sure that you update paths used in the file. In future updates, I'll include the ability to change directories via an interface of some kind.

Now, when you upload an image and then go back, you see the image!